fix(review): establish projected artifact boundary
