+++ src/autoskillit/workspace/_projected_artifact/AGENTS.md
@@ -8,6 +8,7 @@ Projected plugin artifact publication, validation, and launch-lease ownership.
 |------|---------|
 | `__init__.py` | Pure facade for the projected-artifact lifecycle authority |
 | `authority.py` | Cohesive publication, validation, lease handoff, and public constructors |
+| `materialization.py` | Shared projection construction and validation below the authority boundary |
 
 ## Architecture Notes
